requirements.txt
streamlit==1.32.2
chromadb==0.4.24
sentence-transformers==2.6.1
transformers==4.38.2
pypdf==4.2.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.3.0
mistralai==0.4.2
python-dotenv==1.0.1